feat(python): add support for uv (#4440)
This PR picks up on the good work that @theagenticguy started in #4016.

This PR introduces support for the [`uv` python package
manager](https://docs.astral.sh/uv/).

It does not fully deliver the scope that is outlined in #3497, as that
issue is large in scope and relates to the design of `PythonProject`.
This simply introduces `uv` within the existing `PythonProject` design.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: Matt Morris <matt.morris114@gmail.com>
Co-authored-by: Laith Al-Saadoon <lalsaado@amazon.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>