fix: UpgradeDependencies workflow fails in a subproject
`UpgradeDependencies` creates a GitHub workflow to run a task that
it itself creates.

Because the GitHub workflow runs in the repository root by default,
implicit in that assumption is that the project that the task is created
for lives in the repository root as well.

If that is not the case, we get a "no such command" error in the
workflow, an [example of which can be seen
here](https://github.com/aws/aws-cdk-cli/actions/runs/20355620592).

This PR gives projects a conception of a repository-relative directory
and executes the task in the correct directory, while leaving the
repository setup tasks to run in the repository root.

Most likely correct for realistic use cases, but I'm willing to grant
that it depends on the exact structure of the repository and if
the dependencies are managed on a subproject-by-subproject basis
this will still be incorrect.

Not sure how to fix that in projen proper because it does not have
enough information to make this distinction accurately.